Synapse moved to inline processing unit
