+++ libraries/networking/networking-core/Synapse.ipp
@@ -1,11 +1,19 @@
 #pragma once
 
-template <Networking::ConnectionMode mode>
-void Networking::AsyncSynapse<mode>::kill()
-{
-    isAlive = false;
-    proc.join();
+namespace Networking {
+    template<>
+    void Synapse<ConnectionMode::TCP>::listen();
+    template<>
+    void Synapse<ConnectionMode::UDP>::listen();
+
+    template class Synapse<ConnectionMode::UDP>;
+    template class Synapse<ConnectionMode::TCP>;
+
+    template class AsyncSynapse<ConnectionMode::UDP>;
+    template class AsyncSynapse<ConnectionMode::TCP>;
 }
+
+#pragma region SYNAPSE
 
 template <Networking::ConnectionMode mode>
 Networking::Synapse<mode>::Synapse(uint32_t port)
@@ -35,7 +43,6 @@ Networking::Synapse<mode>::Synapse(const ConnectionInfo& connection)
     }
 }
 
-
 template <Networking::ConnectionMode mode>
 Networking::Synapse<mode>::~Synapse()
 {
@@ -44,8 +51,98 @@ Networking::Synapse<mode>::~Synapse()
 }
 
 template <Networking::ConnectionMode mode>
+void Networking::Synapse<mode>::start()
+{
+    isAlive = true;
+
+    listen();
+}
+
+template <Networking::ConnectionMode mode>
+void Networking::Synapse<mode>::send(const AxonMessage& message)
+{
+    if (!isServer)
+        sendTo(message, &socket.conn);
+}
+
+template <Networking::ConnectionMode mode>
+void Networking::Synapse<mode>::sendTo(const AxonMessage& message, const SOCKADDR_IN_T* dest) const
+{
+    send_message<mode>(socket, message.getSerializedBuffer(), message.getSize());
+}
+
+template<>
+void Networking::Synapse<Networking::ConnectionMode::TCP>::listen()
+{
+    // Todo: async support
+
+    char *buffer[SYNAPSE_MESSAGE_MAX_SIZE] = {};
+    SOCKADDR_IN_T host = {};
+
+    SOCKET_T client;
+    do {
+        client = accept_incoming(socket.socket, &host);
+    } while (!CHECK_VALID(client));
+
+    while (isAlive)
+    {
+        int size = recv_tcp_message((char*) buffer, 256, client);
+        if (size > 0)
+        {
+            const char *message = const_cast<char *>(reinterpret_cast<char *> (buffer));
+            AxonMessage message_ = AxonMessage(message, size);
+            onMessageReceived(message_, &host);
+        }
+    }
+}
+
+template<>
+void Networking::Synapse<Networking::ConnectionMode::UDP>::listen()
+{
+    char *buffer[SYNAPSE_MESSAGE_MAX_SIZE] = {};
+    SOCKADDR_IN_T host = {};
+
+    while (isAlive) {
+        int32_t size = recv_udp_message(reinterpret_cast<char *> (buffer), SYNAPSE_MESSAGE_MAX_SIZE, socket.socket,
+                                        &host);
+        if (size > 0) {
+            const char *message = const_cast<char *>(reinterpret_cast<char *> (buffer));
+            AxonMessage message_ = AxonMessage(message, size);
+            onMessageReceived(message_, &host);
+        }
+    }
+}
+
+template <Networking::ConnectionMode mode>
+void Networking::Synapse<mode>::onMessageReceived(const AxonMessage& message, SOCKADDR_IN_T* from)
+{
+    // Notify
+    SynapseMessageReceivedEvent event_ = SynapseMessageReceivedEvent(message, from);
+    events.invoke(&event_);
+}
+
+#pragma endregion
+
+#pragma region ASYNC_SYNAPS
+
+template <Networking::ConnectionMode mode>
+void Networking::AsyncSynapse<mode>::kill()
+{
+    isAlive = false;
+    proc.join();
+}
+
+template <Networking::ConnectionMode mode>
 Networking::AsyncSynapse<mode>::~AsyncSynapse()
 {
     kill();
 }
 
+template <Networking::ConnectionMode mode>
+void Networking::AsyncSynapse<mode>::start()
+{
+    isAlive = true;
+    proc = std::thread(&Networking::AsyncSynapse<mode>::listen, this);
+}
+
+#pragma endregion
