+++ tests/CMakeLists.txt
@@ -2,8 +2,10 @@ include_directories(${CMAKE_HOME_DIRECTORY}/libraries)
 
 include ("CTest")
 
-file(GLOB_RECURSE V_GLOB LIST_DIRECTORIES true "*")
+if (BUILD_TESTS)
+    enable_testing()
 
-message(STATUS "Searching for build config in ${V_GLOB}")
-
-add_all_subdirs()
+    file(GLOB_RECURSE V_GLOB LIST_DIRECTORIES true "*")
+    message(STATUS "Searching for build config in ${V_GLOB}")
+    add_all_subdirs()
+endif()
